it works, and looks better now
