+++ src-tauri/Cargo.toml
@@ -34,8 +34,6 @@ env_logger = "0.9.0"
 strsim = "0.10.0"
 base64 = "0.13.0"
 flate2 = "1.0"
-chrono = "0.4"
-hhmmss = "0.1"
 
 [target.'cfg(windows)'.dependencies]
 windows = { version = "0.38.0",features = [
