+++ src-tauri/src/api/mod.rs
@@ -1,1 +1,2 @@
 pub mod lyric_line;
+pub mod model;
