+++ src-tauri/src/get_lyrics/kugou/mod.rs
@@ -1,0 +1,2 @@
+pub mod get_lyrics;
+pub mod model;
