+++ src-tauri/src/get_lyrics/mod.rs
@@ -1,4 +1,4 @@
 pub mod netease;
-pub mod song_struct;
+pub mod song;
 pub mod lyric_file;
 pub mod kugou;
